requirements.txt
Flask==3.0.1
tabula-py==2.9.3
pandas==2.2.2